dtfile: fix pointer to integer conversion
Use XtArgVal (usually (long)) to convert
(XtPointer) to (int) and back.

This provides safe way to convert and
avoids compiler warning.

+++ cde/programs/dtfile/Desktop.c
@@ -2535,7 +2535,7 @@ PutOnDTCB (
 
    XmUpdateDisplay (w);
 
-   if((int)client_data != 0)
+   if(client_data != 0)
       mbar = XtParent(w);
    else
       mbar = XmGetPostedFromWidget(XtParent(w));
